+++ crates/primitives/Cargo.toml
@@ -11,6 +11,7 @@ workspace = true
 
 [dependencies]
 # alloy
+alloy-consensus.workspace = true
 alloy-eips.workspace = true
 alloy-primitives.workspace = true
 alloy-rpc-types-engine.workspace = true
